requirements.txt
python-telegram-bot==22.6
PyYAML==6.0.2
fastapi==0.115.12
uvicorn==0.34.0
pytest==8.3.5
pytest-asyncio==0.25.3
